+++ CHANGELOG.md
@@ -4,6 +4,16 @@ All notable changes to this project will be documented in this file. This projec
 to [Semantic Versioning] (http://semver.org/). For change log format,
 use [Keep a Changelog] (http://keepachangelog.com/).
 
+## [1.0.0-beta10] - 2024-07-10
+
+### Changed
+
+- INSERT syntax to be conformed to SQL
+
+### Removed
+
+- `Assignment` component on `QueryBuilder` class
+
 ## [1.0.0-beta9] - 2024-03-19
 
 ### Added